third_party/blink/web_tests/crypto/subtle/ecdh/generate-key-expected.txt

Test generating an EC key pair for ECDH.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Generating a key pair...
PASS keyPair.toString() is '[object Object]'
PASS keyPair.publicKey.type is 'public'
PASS keyPair.publicKey.algorithm.name is 'ECDH'
PASS keyPair.publicKey.algorithm.namedCurve is 'P-256'
PASS keyPair.publicKey.usages is []
PASS keyPair.privateKey.type is 'private'
PASS keyPair.privateKey.algorithm.name is 'ECDH'
PASS keyPair.privateKey.algorithm.namedCurve is 'P-256'
PASS keyPair.privateKey.usages is ["deriveBits"]

Testing that custom attributes on keys survive garbage collection...
PASS keyPair.publicKey.foo is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
